+++ server/src/modules/logger/logger.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 
-import { LoggerService } from "./logger.service";
+import { LoggerService } from "./services";
 
 @Module({
 	providers: [LoggerService],
